Relocate feed to `FundedProjects` controller.
