+++ app/controllers/home_controller.rb
@@ -3,13 +3,4 @@ class HomeController < ApplicationController
     @chapters = Chapter.active.visitable.for_display.all
     @projects = Project.recent_winners.limit(15).all
   end
-
-  def feed
-    project_filter = ProjectFilter.new(Project.includes(:photos).recent_winners)
-    @projects = project_filter.page(params[:page], 50).result
-
-    respond_to do |format|
-      format.xml
-    end
-  end
 end
