+++ spec/controllers/home_controller_spec.rb
@@ -27,17 +27,4 @@ describe HomeController do
       response.body.should have_selector('section.awesome-chapters', :text => chapter.name)
     end
   end
-
-  context 'getting an Atom feed of winning projects' do
-    let!(:chapter) { FactoryGirl.create(:chapter, country: 'United States') }
-
-    before do
-      get :feed, format: :xml
-    end
-
-    it 'should return parseable XML' do
-      response.content_type.should eq('application/xml')
-      lambda { Nokogiri::XML(response.body) }.should_not raise_error
-    end
-  end
 end
